Change forms calculations: * Make both calculations return integers * Change validations with new logic (trueNetIncome=netIncome-Obligations) * Add customExceptions

+++ src/main/java/com/team5/morgage/controllers/CalculatorController.java
@@ -25,7 +25,7 @@ public class CalculatorController {
     }
 
     @PostMapping("/maxLoan")
-    public ResponseEntity<Float> calculateMaxLoan(@RequestBody @Valid MaxLoanRequest maxLoanRequest) {
+    public ResponseEntity<Integer> calculateMaxLoan(@RequestBody @Valid MaxLoanRequest maxLoanRequest) {
         return new ResponseEntity<>(calculatorService.calculateMaxLoan(maxLoanRequest), HttpStatus.CREATED);
     }
